created jsonnetapidoc handler
