+++ src/OpenRasta.Plugins.Hydra/Configuration/HydraResourceModel.cs
@@ -12,5 +12,6 @@ namespace OpenRasta.Plugins.Hydra.Configuration
     public Func<object,SerializationOptions, Stream,Task> SerializeFunc { get; set; }
     public Vocabulary Vocabulary { get; set; }
     public List<Operation> SupportedOperations { get; set; } = new List<Operation>();
+    public Class ApiClass { get; set; }
   }
 }
