+++ src/OpenRasta.Plugins.Hydra/Internal/Serialization/JsonNet/HydraUriModelConverter.cs
